+++ .github/workflows/install.sh
@@ -11,7 +11,7 @@ echo "Installing package and requirements for ${flavor}"
 
 if [[ "${flavor}" = "testing" ]]; then
   poetry run pip install wradlib --no-deps
-  poetry install --no-interaction --extras=sql --extras=export --extras=restapi --extras=explorer
+  poetry install --no-interaction --extras=sql --extras=export --extras=restapi --extras=explorer --extras=bufr
 
 elif [[ "${flavor}" = "docs" ]]; then
   poetry install --no-interaction --extras=docs